refactor!: remove utility map exports
