+++ src/dom/at-tabindex.ts
@@ -1,5 +1,5 @@
 import {DomWatcher} from '../dom/dom-watcher';
-import {func} from '../func/func';
+import * as func from '../func/func';
 import * as dom from '../dom/dom';
 
 /**
