+++ src/dom/dom.spec.ts
@@ -1,5 +1,5 @@
 import * as is from '../is/is';
-import {dom} from './dom';
+import * as dom from './dom';
 import test from 'ava';
 
 test('Dom is defined', t => {
