+++ src/dom/dom.ts
@@ -1298,56 +1298,3 @@ export function addDelegatedListener(
     el.removeEventListener(type, handler);
   };
 }
-
-/**
- * Degu DOM utility functions.
- */
-export const dom = {
-  addDelegatedListener,
-  addStyles,
-  addStylesToPage,
-  appendAfter,
-  aspect,
-  copyBase64Image,
-  createElementFromString,
-  deleteImage,
-  disableScrolling,
-  event,
-  fetchAndMakeImage,
-  flushVideos,
-  forceFocus,
-  forceVOFocus,
-  getAllTextNodes,
-  getComputedStyle,
-  getElementScrolledPercent,
-  getFocusableElements,
-  getScrollElement,
-  getScrollTop,
-  getScrollYAtPercent,
-  getStyle,
-  isDisplayNone,
-  isDisplayNoneWithAncestors,
-  isOverlapping,
-  isVisibleOnScreen,
-  makeBase64ImageFromBlob,
-  makeImageFromBlob,
-  pauseAllVideosInElement,
-  playAllVideosInElement,
-  removeClassByPrefix,
-  removeElement,
-  removeAllChildren,
-  resetAllVideosInElement,
-  resetForceFocus,
-  runAfterNotTopOfScreen,
-  runAfterWindowLoad,
-  setCssVariable,
-  setCssVariables,
-  setFocusableElementsToTabIndex,
-  testDescendant,
-  testVideoIsPlaying,
-  unflushVideos,
-  unorphan,
-  unsetFocusableElementsToTabIndex,
-  whenVideosLoaded,
-  whenVideosLoadedInElement,
-};
