+++ src/mathf/color.spec.ts
@@ -1,4 +1,4 @@
-import {color} from './color';
+import * as color from './color';
 import test from 'ava';
 
 test('hexToRgbNormalized', t => {
