+++ src/mathf/mathf.ts
@@ -1447,73 +1447,3 @@ export function remap(
 export function sum(values: number[]) {
   return values.reduce((result, value) => result + value, 0);
 }
-
-/**
- * Degu Math utility functions.
- */
-export const mathf = {
-  absZero,
-  angleDegree,
-  angleDistanceDegree,
-  angleDistanceRadian,
-  angleRadians,
-  aspectRatio,
-  asymptoticAverage,
-  boxCollision,
-  calculate2dPointRotation,
-  calculate2dPointRotationWithDistance,
-  calculateBackgroundContain,
-  calculateBackgroundCover,
-  calculateCenterOffset,
-  ceilToPrecision,
-  childProgress,
-  clamp,
-  clamp01,
-  clampAsPercent,
-  clampAsProgress,
-  collisionCircVsBox,
-  collisionPointVersusConvexPolygon,
-  damp,
-  degreeToRadian,
-  direction,
-  distance,
-  ease,
-  fixDigits,
-  flipCoin,
-  floorToPrecision,
-  generateRandomPoints,
-  getProgressInRangeByValue,
-  getRandomFloat,
-  getRandomInt,
-  getUniqueRandomInt,
-  getValueInRangeByProgress,
-  int,
-  int0,
-  interpolateRange,
-  inverseLerp,
-  inverseProgress,
-  isBetween,
-  lerp,
-  lerpEase,
-  mix,
-  normalize,
-  radianToDegree,
-  remap,
-  resizeDimensionalBoxToHeight,
-  resizeDimensionalBoxToWidth,
-  roundToPrecision,
-  scaleX,
-  scaleY,
-  sigmoid,
-  sinNormalized,
-  smoothStart2,
-  smoothStart3,
-  smoothStep,
-  smoothStep2,
-  smoothStop2,
-  smoothStop3,
-  step,
-  sum,
-  toFixed,
-  wrap,
-};
